// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/counterOfferResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CounterOfferResponse {
  counterOfferId?: number;
  quoteId?: number;
  driverId?: number;
  proposedPrice?: number;
  message?: string;
  status?: string;
  createdAt?: string;
  respondedAt?: string;
}