// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/quoteListResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface QuoteListResponse {
  quoteId?: number;
  quotePublicId?: string;
  truckId?: number;
  originAddress?: string;
  destinationAddress?: string;
  distanceKm?: number;
  vehicleType?: string;
  vehicleBodyType?: string;
  cargoName?: string;
  desiredPrice?: number;
  finalPrice?: number;
  status?: string;
  createdAt?: string;
}